requirements.txt
DI-engine>=0.5.3
gymnasium[atari]==0.28.0
numpy>=1.24.1,<2
pympler
minigrid
moviepy
pytest
line_profiler
xxhash
einops
